crawler simplification
